feat: 검색 예외 알림과 화면 유틸 분리
